fix active example
