+++ digital_twins/incubator-TeSSLa-active-control/lifecycle/execute.py
@@ -1,63 +1,10 @@
-import subprocess, os, time, sys
+import subprocess, os, time, sys, traceback
 incubator_location = os.getenv("INCUBATOR_PATH")
-tessla_location = os.getenv("TESSLA_PATH")
-lifecycle_location = os.getenv("LIFECYCLE_PATH")
 sys.path.append(os.path.join(f"{incubator_location}"))
-sys.path.append(os.path.join(f"{tessla_location}"))
 from threading import Thread, Event
 from incubator.communication.server.rabbitmq import Rabbitmq
 from incubator.config.config import load_config
 from digital_twin.communication.rabbitmq_protocol import ROUTING_KEY_ENERGY_SAVER_ENABLE, ROUTING_KEY_ENERGY_SAVER_STATUS, ROUTING_KEY_LIDOPEN
-
-
-def makeTelegrafConfig(template_path, dest_path):
-    config = load_config(f"{incubator_location}/simulation.conf")
-    try:
-        with open(template_path, 'r') as file:
-            content = file.read()
-
-        content = content.replace('<AMQP_HOST>', config.get_string('rabbitmq.ip'))
-        content = content.replace('<AMQP_PORT>', config.get_string('rabbitmq.port'))
-        content = content.replace('<AMQP_VHOST>', config.get_string('rabbitmq.vhost'))
-        content = content.replace('<AMQP_USER>', config.get_string('rabbitmq.username'))
-        content = content.replace('<AMQP_PASS>', config.get_string('rabbitmq.password'))
-        content = content.replace('<AMQP_EXCHANGE>', config.get_string('rabbitmq.exchange'))
-
-        with open(dest_path, 'w') as file:
-            file.write(content)
-
-    except Exception as e:
-        print(f"An error occurred while generating the telegraf config file: {e}")
-        sys.exit(1)
-
-
-def startTessla():
-    tesslaProcess = subprocess.Popen(f"cd {tessla_location}/; exec ./TesslaTelegrafConnector -i safe-operation.tessla -c telegraf.conf -r 2>&1 | tee /tmp/tessla.log", shell=True, cwd=os.getcwd(), stderr=subprocess.PIPE, stdout=subprocess.PIPE, text=True)
-    output = ""
-    while True:
-        line = tesslaProcess.stdout.readline()
-        output += line
-        if line == "" and tesslaProcess.poll() is not None:
-            print("Error starting TeSSLa-Telegraf-Connector:")
-            print(output)
-            print(tesslaProcess.stderr.readlines())
-            sys.exit(1)
-        if "Start listening..." in line :
-            print("Started TeSSLa-Telegraf-Connector")
-            break
-
-    return tesslaProcess
-
-
-def startTelegraf(conf_path):
-    print("Starting Telegraf")
-    telegrafProcess = subprocess.Popen(f"exec telegraf --config {conf_path}", shell=True, cwd=os.getcwd(), stderr=subprocess.PIPE, stdout=subprocess.PIPE, text=True)
-    time.sleep(5)
-    if telegrafProcess.poll() is not None:
-        print("Failed to start telegraf server")
-        print(telegrafProcess.stdout.readlines())
-        sys.exit(1)
-    return telegrafProcess
 
 
 def startRabbitMQ(message_callback):
@@ -67,6 +14,7 @@ def startRabbitMQ(message_callback):
     print("Connected to rabbitmq server.")
     rabbitMq.subscribe(ROUTING_KEY_ENERGY_SAVER_STATUS, handleMessage)
     rabbitMq.subscribe(ROUTING_KEY_LIDOPEN, handleMessage)
+    rabbitMq.subscribe("incubator.update.closed_loop_controller.parameters", handleMessage)
     return rabbitMq
 
 
@@ -117,38 +65,34 @@ def runScenario(event):
 
 if __name__ == "__main__":
 
-    tesslaProcess = None
-    telegrafProcess = None
     rabbitMq = None
     incubatorProcess = None
     scenario_thread = None
+    event = Event()
     try:
-        makeTelegrafConfig(f"{lifecycle_location}/../telegraf.conf", f"{tessla_location}/telegraf.conf")
-        tesslaProcess = startTessla()
-        telegrafProcess = startTelegraf(f"{tessla_location}/telegraf.conf")
         # Start incubator
         incubatorProcess = startIncubator()
         message = {}
         message["anomaly"] = "!anomaly"
         message["energy_saving"] = "!energy_saving"
-        message["alert"] = "unknown"
+        message["temp"] = "unknown"
 
         # setup RabbitMQ
         def handleMessage(channel, method, properties, body_json):
+            #print(f"Channel: {channel}. Method: {method}. Properties: {properties}. Body: {body_json}.")
             if "lid_open" in body_json:
                 message["anomaly"] = "anomaly" if body_json["lid_open"] else "!anomaly"
             elif "energy_saver_on" in body_json:
                 message["energy_saving"] = "energy_saving" if body_json["energy_saver_on"] else "!energy_saving"
-            elif "alert" in body_json:
-                message["alert"] = "alert" if body_json["alert"] else "unknown"
-                print(f"Message from TeSSLa: {body_json['alert']}")
+            elif "temperature_desired" in body_json:
+                message["temp"] = int(body_json["temperature_desired"])
+                print(f"Message from TeSSLa: {body_json}")
             states = f"{message['anomaly']} & {message['energy_saving']}"
-            print(f"State: {states}, verdict: {message['alert']}")
+            print(f"State: {states}, Desired Temp: {message['temp']}")
 
             if event.is_set():
                 rabbitMq.close()
         rabbitMq = startRabbitMQ(handleMessage)
-        event = Event()
         scenario_thread = Thread(target=runScenario, args=[event])
         # Wait to ensure incubator running
         # time.sleep(1)
@@ -158,18 +102,11 @@ if __name__ == "__main__":
     except KeyboardInterrupt:
         event.set()
         print("Stopping simulation...")
-    except:
+    except Exception as e:
         event.set()
-        print("An error has occurred")
+        print("An error has occurred:")
+        traceback.print_exc()
     finally:
-        if tesslaProcess:
-            print(f"Stopping TeSSLa with pid: {tesslaProcess.pid}")
-            tesslaProcess.kill()
-            tesslaProcess.wait()
-        if telegrafProcess:
-            print(f"Stopping telegraf with pid: {telegrafProcess.pid}")
-            telegrafProcess.kill()
-            telegrafProcess.wait()
         if rabbitMq and not event.is_set():
             rabbitMq.close()
         if incubatorProcess:
