+++ digital_twins/incubator-TeSSLa-active-control/telegraf.conf
@@ -1,7 +1,7 @@
 
 [agent]
-interval = "1s"
-flush_interval = "1s"
+	interval = "1s"
+	flush_interval = "1s"
 
 [[inputs.amqp_consumer]]
 	brokers = ["amqp://<AMQP_HOST>:<AMQP_PORT>/<AMQP_VHOST>"]
@@ -34,8 +34,8 @@ flush_interval = "1s"
 	exchange = "<AMQP_EXCHANGE>"
 	exchange_durability = "transient"
 	routing_key = "incubator.update.closed_loop_controller.parameters"
-	namepass = ["result"]
 	data_format = "json"
+	namepass = ["result"]
 	use_batch_format = false
 	json_transformation = '{"temperature_desired" : fields.value ? 21 : 35}'
